[GEOS-8385] Update MongoDB integration with App-Schema tests to use relative mappings
